third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/event-is-fired.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('body_sensor_location'))
    .then(characteristic => {
      return assert_promise_resolves_after_event(characteristic,
                                                 'readValue',
                                                 'characteristicvaluechanged');
    }).then(results => {
      let read_value = results[0].buffer;
      let event_value = results[1].buffer;
      assert_equals(event_value, read_value);
      assert_array_equals(event_value, read_value);
    });
}, 'Reading a characteristic should fire an event.');
</script>
